+++ driver/src/main/com/mongodb/rx/client/FindObservable.java
@@ -17,6 +17,7 @@
 package com.mongodb.rx.client;
 
 import com.mongodb.CursorType;
+import com.mongodb.client.model.Collation;
 import org.bson.conversions.Bson;
 import rx.Observable;
 
@@ -151,4 +152,15 @@ public interface FindObservable<TResult> extends MongoObservable<TResult> {
      * @return this
      */
     FindObservable<TResult> cursorType(CursorType cursorType);
+
+    /**
+     * Sets the collation options
+     *
+     * <p>A null value represents the server default.</p>
+     * @param collation the collation options to use
+     * @return this
+     * @since 1.3
+     * @mongodb.server.release 3.4
+     */
+    FindObservable<TResult> collation(Collation collation);
 }