+++ driver/src/test/unit/com/mongodb/rx/client/FindObservableSpecification.groovy
@@ -23,6 +23,7 @@ import com.mongodb.async.AsyncBatchCursor
 import com.mongodb.async.client.FindIterable
 import com.mongodb.async.client.FindIterableImpl
 import com.mongodb.async.client.MongoIterable
+import com.mongodb.client.model.Collation
 import com.mongodb.client.model.FindOptions
 import com.mongodb.operation.FindOperation
 import org.bson.BsonDocument
@@ -47,10 +48,11 @@ class FindObservableSpecification extends Specification {
 
     def namespace = new MongoNamespace('db', 'coll')
     def codecRegistry = fromProviders([new DocumentCodecProvider(), new BsonValueCodecProvider(), new ValueCodecProvider()])
+    def collation = Collation.builder().locale('en').build()
 
     def 'should have the same methods as the wrapped FindIterable'() {
         given:
-        def wrapped = (FindIterable.methods*.name - MongoIterable.methods*.name).sort()  - 'collation'
+        def wrapped = (FindIterable.methods*.name - MongoIterable.methods*.name).sort()
         def local = (FindObservable.methods*.name - MongoObservable.methods*.name - 'first' - 'batchSize').sort()
 
         expect:
@@ -72,6 +74,7 @@ class FindObservableSpecification extends Specification {
                 .oplogReplay(false)
                 .noCursorTimeout(false)
                 .partial(false)
+                .collation(collation)
         def wrapped = new FindIterableImpl<Document, Document>(namespace, Document, Document, codecRegistry, secondary(),
                 ReadConcern.DEFAULT, executor, new Document('filter', 1), findOptions)
         def findObservable = new FindObservableImpl<Document>(wrapped, new ObservableHelper.NoopObservableAdapter())
@@ -94,10 +97,12 @@ class FindObservableSpecification extends Specification {
                 .skip(10)
                 .cursorType(CursorType.NonTailable)
                 .slaveOk(true)
+                .collation(collation)
         )
         readPreference == secondary()
 
         when: 'overriding initial options'
+        def frenchCollation = Collation.builder().locale('fr').build()
         subscriber = new TestSubscriber(100)
         findObservable.filter(new Document('filter', 2))
                 .sort(new Document('sort', 2))
@@ -111,6 +116,7 @@ class FindObservableSpecification extends Specification {
                 .oplogReplay(true)
                 .noCursorTimeout(true)
                 .partial(true)
+                .collation(frenchCollation)
                 .subscribe(subscriber)
 
         operation = executor.getReadOperation() as FindOperation<Document>
@@ -131,6 +137,7 @@ class FindObservableSpecification extends Specification {
                 .noCursorTimeout(true)
                 .partial(true)
                 .slaveOk(true)
+                .collation(frenchCollation)
         )
     }