+++ driver/src/test/unit/com/mongodb/rx/client/MapReduceObservableSpecification.groovy
@@ -22,6 +22,7 @@ import com.mongodb.WriteConcern
 import com.mongodb.async.client.MapReduceIterable
 import com.mongodb.async.client.MapReduceIterableImpl
 import com.mongodb.async.client.MongoIterable
+import com.mongodb.client.model.Collation
 import com.mongodb.client.model.MapReduceAction
 import com.mongodb.operation.FindOperation
 import com.mongodb.operation.MapReduceToCollectionOperation
@@ -48,10 +49,11 @@ class MapReduceObservableSpecification extends Specification {
     def namespace = new MongoNamespace('db', 'coll')
     def codecRegistry = fromProviders([new ValueCodecProvider(), new DocumentCodecProvider(), new BsonValueCodecProvider()])
     def readPreference = secondary()
+    def collation = Collation.builder().locale('en').build()
 
     def 'should have the same methods as the wrapped MapReduceIterable'() {
         given:
-        def wrapped = (MapReduceIterable.methods*.name - MongoIterable.methods*.name).sort() - 'collation'
+        def wrapped = (MapReduceIterable.methods*.name - MongoIterable.methods*.name).sort()
         def local = (MapReduceObservable.methods*.name - MongoObservable.methods*.name - 'batchSize').sort()
 
         expect:
@@ -86,6 +88,7 @@ class MapReduceObservableSpecification extends Specification {
                 .scope(new Document('scope', 1))
                 .sort(new Document('sort', 1))
                 .verbose(false)
+                .collation(collation)
                 .subscribe(subscriber)
 
         operation = executor.getReadOperation() as MapReduceWithInlineResultsOperation<Document>
@@ -100,6 +103,7 @@ class MapReduceObservableSpecification extends Specification {
                 .scope(new BsonDocument('scope', new BsonInt32(1)))
                 .sort(new BsonDocument('sort', new BsonInt32(1)))
                 .verbose(false)
+                .collation(collation)
         )
     }
 
@@ -126,6 +130,7 @@ class MapReduceObservableSpecification extends Specification {
                 .action(MapReduceAction.MERGE)
                 .sharded(true)
                 .jsMode(true)
+                .collation(collation)
                 .subscribe(subscriber)
 
         def operation = executor.getWriteOperation() as MapReduceToCollectionOperation
@@ -143,6 +148,7 @@ class MapReduceObservableSpecification extends Specification {
                 .action(MapReduceAction.MERGE.getValue())
                 .jsMode(true)
                 .sharded(true)
+                .collation(collation)
 
         then: 'should use the overrides'
         expect operation, isTheSameAs(expectedOperation)